dmp-service/requirements.txt
confluent-kafka
psycopg2-binary
pyyaml